# frontend/src/services/openapi.yml
openapi: 3.0.2
info:
  title: Microservices API
  description: API per l'autenticazione, gestione magazzino, ordini, pagamenti e spedizioni
  version: 1.0.0
  contact:
    name: Daniele Zoccali
    email: daniele.zoccali@gmail.com
servers:
  - url: https://microservices-alb-216051693.us-east-1.elb.amazonaws.com
    description: Server principale (ALB)
  - url: https://localhost:3002
    description: Server di sviluppo per Warehouse
  - url: https://localhost:3005
    description: Server di sviluppo per Payments
security:
  - bearerAuth: [] #"Per default, tutte le rotte richiedono un token JWT nel campo Authorization: Bearer <token>. "
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500.
        error:
          description: "Descrizione dell'errore; può essere una stringa o un oggetto."
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500
        request_id:
          type: string
          maxLength: 100
          description: "Identificativo della richiesta (se presente)."
          pattern: '^[0-9a-f]{32}$'
          example: "fd723645da950a54df847147b19f1108"
        success:
          type: boolean
          description: "Flag che, in alcuni casi, è restituito anche in risposta di errore."
    NoRoutes:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500.
        request_id:
          type: string
          maxLength: 100
          description: "Identificativo della richiesta (se presente)."
          pattern: '^[0-9a-f]{32}$'
          example: "fd723645da950a54df847147b19f1108"
...
    Product:
      type: object
      additionalProperties: false
      required:
        - _id
        - name
        - price
        - image
        - qty
      properties:
        _id:
          type: string
          maxLength: 24
          pattern: '^[0-9a-f]{24}$'  # stringa esadecimale di 24 caratteri
        name:
          type: string
          maxLength: 100
          pattern: '^.{1,100}$'  # qualsiasi stringa da 1 a 100 caratteri
        price:
          type: number
          minimum: 0
          maximum: 9999999
        image:
          type: array
          items:
            type: string
            format: uri
            maxLength: 2048
            pattern: '^(https?:\/\/).+$'
          maxItems: 100
        qty:
          type: integer
          minimum: 1
          maximum: 9999999
        __v:
          type: integer
          minimum: 0
...
    ErrorOnlyMessage:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 500
          pattern: '^.{1,500}$' # accetta qualsiasi stringa con almeno 1 carattere fino a 500.
...
paths:
...
  /admin/warehouse/products:
    servers:
    - url: https://localhost:3002
      description: Server di sviluppo per Warehouse
    get:
      summary: Ottieni tutti i prodotti nel magazzino
      tags:
        - Warehouse
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista di prodotti nel magazzino
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
                maxItems: 10
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoRoutes'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

#OK happy path in teoria
  /admin/warehouse/product/{productId}/quantity:
    servers:
    - url: https://localhost:3002
      description: Server di sviluppo per Warehouse
    patch:
      summary: Aggiorna quantità prodotto
      tags:
        - Warehouse
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$' #stringa esadecimale di 24 caratteri
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - change
              properties:
                change:
                  type: integer
                  format: int32
                  minimum: -9999999
                  maximum: 9999999
            examples:
              updateQty:
                summary: Update valido      
                value:
                  change: 1
      responses:
        '200':
          description: Quantità aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Dati non validi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Prodotto non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/html:
              schema:
                type: string
                maxLength: 4096
                pattern: '^<html>[\\s\\S]{1,4090}</html>$'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/warehouse/product/{productId}:
    servers:
    - url: https://localhost:3002
      description: Server di sviluppo per Warehouse
    get:
      summary: Ottieni dettagli di un prodotto
      tags:
        - Warehouse
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$' #stringa esadecimale di 24 caratteri
      responses:
        '200':
          description: Dettagli del prodotto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Prodotto non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/html:
              schema:
                type: string
                maxLength: 4096
                pattern: '^<html>[\\s\\S]{1,4090}</html>$'
                
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Elimina un prodotto
      tags:
        - Warehouse
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '^[0-9a-f]{24}$' #stringa esadecimale di 24 caratteri
      responses:
        '200':
          description: Elimanto con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOnlyMessage'
              examples:
                validDelete:
                  summary: Eliminazione valida - Happy path
                  value:
                    message: "Prodotto eliminato con successo"
        '401':
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Prodotto non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoRoutes'
            text/html:
              schema:
                type: string
                maxLength: 4096
                pattern: '^<html>[\\s\\S]{1,4090}</html>$'
        '406':
          description: Not Acceptable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Errore non previsto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'







#Ok HAPPY PATH
...